requirements.txt
asgiref==3.8.1
certifi==2025.1.31
cloudinary==1.42.2
dj-database-url==2.3.0
Django==5.1.6
django-filter==25.1
djangorestframework==3.15.2
Markdown==3.7
pillow==11.1.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==2.3.0
